Add show_add_existing and show_add_new helpers.
Move links_for_associations back to nested.rb. The before_filter :require_user is not called before the first call to self.links_for_associations on the refresh and with authlogic the current_user is not defined properly. I can pretty_inspect it but as soon as I try to access one of it's methods from inside authorized_for_read? I get an error, the object seems in tact for an object comparison but that is all. The whole mechanism of Model#current_user seems a bit timing dependent. Until I figure it out it is going back into nested.rb